+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.18.0
 - add missing export to json output for `extract` command
 - adds analysis of readability and writability of fields.
+- fix change compatibility check for `dynamic` and `Object?` parameters
 
 ## Version 0.17.0
 - bump dart SDK requirements to >=3.0.0